+++ realm/realm-library/src/main/cpp/io_realm_internal_TableQuery.cpp
@@ -1920,6 +1920,6 @@ JNIEXPORT jlong JNICALL Java_io_realm_internal_TableQuery_nativeGetFinalizerPtr
   (JNIEnv *, jclass)
 {
     TR_ENTER()
-    return static_cast<jlong>(reinterpret_cast<uintptr_t>(&finalize_table_query));
+    return reinterpret_cast<jlong>(&finalize_table_query);
 }
